+++ MSMetaEnhancer/libs/services/__init__.py
@@ -1,6 +1,6 @@
 from MSMetaEnhancer.libs.services.CIR import CIR
 from MSMetaEnhancer.libs.services.CTS import CTS
 from MSMetaEnhancer.libs.services.NLM import NLM
-from MSMetaEnhancer.libs.services.PubChem import PubChem
+from MSMetaEnhancer.libs.services.IDSM import IDSM
 
-__all__ = ['PubChem', 'CTS', 'CIR', 'NLM']
+__all__ = ['IDSM', 'CTS', 'CIR', 'NLM']