+++ MSMetaEnhancer/libs/converters/compute/__init__.py
@@ -1,0 +1,3 @@
+from MSMetaEnhancer.libs.converters.compute.RDKit import RDKit
+
+__all__ = ['RDKit']