+++ CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [dev] - unreleased
 ### Added
+* reintroduced PubChem service using direct REST web interface
 ### Changed
 * renamed PubChem service to IDSM to avoid confusion
 ### Removed